requirements.txt
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "3.12"
django-vite==2.1.3 ; python_version >= "3.10" and python_version < "3.12"
django==4.2.3 ; python_version >= "3.10" and python_version < "3.12"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "3.12"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "3.12"
python-decouple==3.8 ; python_version >= "3.10" and python_version < "3.12"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.12"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "3.12"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "3.12"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "win32"
whitenoise==6.5.0 ; python_version >= "3.10" and python_version < "3.12"
